requirements.txt
alembic==1.8.1
aniso8601==9.0.1
asyncpg==0.26.0
attrs==22.1.0
boto3==1.24.88
botocore==1.27.88
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.1
ecdsa==0.18.0
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.9.0
Inject==4.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.10.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.4
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.1.2
zipp==3.8.1
